Add shared quality efficiency criteria
