+++ tasks/dataset.toml
@@ -11,49 +11,49 @@ name = "Tempo"
 
 [[tasks]]
 name = "tempo/transfer-with-memo-docs"
-digest = "sha256:f20b3e0e90b81549ff6b3995c1986e194dac3403611549426fc178d01fcf9bc4"
+digest = "sha256:d50d0da7dd18e3d48cc28487792d5202651bfd525bd9ae272c3c2b2a34ecfd1e"
 
 [[tasks]]
 name = "tempo/transfer-with-memo-mcp"
-digest = "sha256:2f04ae7cea2d878807bf2b8b83cfdfae6b04ca210ac7205552d01b946f0a5a4f"
+digest = "sha256:cd733540eb8bcf14762f5057ae87c110eea14e31a9ec2ad8a381a522f48e350a"
 
 [[tasks]]
 name = "tempo/transfer-with-memo-fee-payer-docs"
-digest = "sha256:5440d62b3fb5f764f85e800ee03556ca1284dd16d7b19a854c11f13d1a9e339c"
+digest = "sha256:5cd6e98df033072c68ae80cd41d0b725530bbb0615a8f285f53e6b050081ab5b"
 
 [[tasks]]
 name = "tempo/transfer-with-memo-fee-payer-mcp"
-digest = "sha256:04e2921c5c1c3135e6052c5dc8d7a2491b644ba7ef57c106012d472aeb42bd41"
+digest = "sha256:0322f37f208d6bdfc03df8b5ae99277eb27630b5b771149507c4d50c468bce30"
 
 [[tasks]]
 name = "tempo/set-fee-token-docs"
-digest = "sha256:99a6b439d8df6aceda3a9bb590fdf089d64cd6ae7ab12c1c5502e5a1f789c106"
+digest = "sha256:aa70ccf5fe07e546bc415395b2e87150985b5aa2f2b27a35192be3117e135563"
 
 [[tasks]]
 name = "tempo/set-fee-token-mcp"
-digest = "sha256:d2cdc3b09f3817f622b90fa4b11ab381dce295b9806e933ff830d9ec0be96f38"
+digest = "sha256:fe0cdb230f4630eada05b950ac7ace6e347c880a46a1e8ee9eb6d81000509a26"
 
 [[tasks]]
 name = "tempo/create-stablecoin-with-policy-docs"
-digest = "sha256:a7189eba0ec2bba9d24cd76b926565d5b87d0bd40a4cc057248a3446b56d498e"
+digest = "sha256:03ad34b9321747df339c71b74d1987b68abb892d01eeda754c47fe6aea6cb79d"
 
 [[tasks]]
 name = "tempo/create-stablecoin-with-policy-mcp"
-digest = "sha256:5db127ba6e53ab68cf87cd6f034f91cb752f9a0b354e03cdd5466041468fd1dc"
+digest = "sha256:ac899df1920d5f2888b858b224413dff086ba08e9680e273db04055abd6dec8d"
 
 [[tasks]]
 name = "tempo/faucet-funded-transfer-docs"
-digest = "sha256:bb87531d1b9e97acb2c4b63ffb510e437c987f86260da28e824d1d7cf2eb8567"
+digest = "sha256:d66ff3f56deeba707d5ccb8417effa3522b30a3e3d131735f7ff714187a911d1"
 
 [[tasks]]
 name = "tempo/faucet-funded-transfer-mcp"
-digest = "sha256:bd009945a1c6eda9d61dd08cc6812ea837beba993ceaf583bb53aa1b706cba01"
+digest = "sha256:17cb071a40b66c80b11b28d57fa0c6bdc36d08a014e4d20aa9c5ab7eb8c68440"
 
 [[tasks]]
 name = "tempo/stablecoin-dex-swap-docs"
-digest = "sha256:2ff8c305a400d1995b16f562bed5469c508bbbe03a6c9b1359d1e1ed4e9d8825"
+digest = "sha256:1bbf9831574c8aa7d58bf72c57bc07f1cd514472581330645addf6d581a54331"
 
 [[tasks]]
 name = "tempo/stablecoin-dex-swap-mcp"
-digest = "sha256:cf382040be5231b9eab24ab8500ca510ddebc6760b50a312a0cfdbe68d5ba12b"
+digest = "sha256:9d4bec9b25c79c6131119e37413c22213fed2d1d33372c0635f6809e773aea9a"
